requirements.txt
# Supabase Client
supabase>=2.0.0

# Excel Processing
openpyxl>=3.1.0
pandas>=2.0.0

# Configuration
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.0.0

# Testing
pytest>=7.4.0
hypothesis>=6.90.0

# Progress Bars
tqdm>=4.66.0

# Additional Utilities
python-dateutil>=2.8.2
backoff>=2.2.0
